Move constants to system.
