+++ packages/compiler/src/Compiler.ts
@@ -4,15 +4,9 @@ import ejs, { TemplateFunction } from 'ejs';
 import prettier, { BuiltInParserName } from 'prettier';
 import camelCase from 'lodash/camelCase';
 import { Path, PortablePath } from '@boost/common';
-import { createMixins, DEPTHS } from '@aesthetic/system';
-import BrandLoader from './BrandLoader';
-import LanguageLoader from './LanguageLoader';
-import ThemesLoader from './ThemesLoader';
-import SystemDesign from './SystemDesign';
-import SystemTheme from './SystemTheme';
-import WebPlatform from './platforms/Web';
-import { FormatType, SystemOptions, PlatformType, MixinsTemplate, ThemesConfigFile } from './types';
 import {
+  createMixins,
+  DEPTHS,
   BORDER_SIZES,
   BREAKPOINT_SIZES,
   HEADING_SIZES,
@@ -21,7 +15,14 @@ import {
   SHADOW_SIZES,
   SPACING_SIZES,
   TEXT_SIZES,
-} from './constants';
+} from '@aesthetic/system';
+import BrandLoader from './BrandLoader';
+import LanguageLoader from './LanguageLoader';
+import ThemesLoader from './ThemesLoader';
+import SystemDesign from './SystemDesign';
+import SystemTheme from './SystemTheme';
+import WebPlatform from './platforms/Web';
+import { FormatType, SystemOptions, PlatformType, MixinsTemplate, ThemesConfigFile } from './types';
 
 type Platform = WebPlatform;
