+++ packages/compiler/src/ThemesLoader.ts
@@ -1,9 +1,9 @@
 import { deepMerge } from '@aesthetic/utils';
-import { DeepPartial, ContrastLevel, ColorScheme, Hexcode } from '@aesthetic/system';
+import { DeepPartial, ContrastLevel, ColorScheme, Hexcode, SHADE_RANGES } from '@aesthetic/system';
 import { Path } from '@boost/common';
 import optimal, { number, object, ObjectOf, Schema, shape, string, union } from 'optimal';
 import Loader from './Loader';
-import { SHADE_RANGES, THEMES_FILE } from './constants';
+import { THEMES_FILE } from './constants';
 import {
   ThemesConfigFile,
   ThemeConfig,
