+++ packages/system/src/constants.ts
@@ -1,4 +1,18 @@
-import { ElevationType } from './types';
+import {
+  BorderSize,
+  BreakpointSize,
+  ElevationType,
+  HeadingSize,
+  ShadowSize,
+  SpacingSize,
+  TextSize,
+  PaletteType,
+  ColorShade,
+} from './types';
+
+export const BORDER_SIZES: BorderSize[] = ['sm', 'df', 'lg'];
+
+export const BREAKPOINT_SIZES: BreakpointSize[] = ['xs', 'sm', 'md', 'lg', 'xl'];
 
 export const DEPTHS: Record<ElevationType, number> = {
   content: 100, // xs
@@ -11,3 +25,38 @@ export const DEPTHS: Record<ElevationType, number> = {
   menu: 1600, // md
   tooltip: 1700, // sm
 };
+
+export const HEADING_SIZES: HeadingSize[] = ['l1', 'l2', 'l3', 'l4', 'l5', 'l6'];
+
+export const SHADOW_SIZES: ShadowSize[] = ['xs', 'sm', 'md', 'lg', 'xl'];
+
+export const SPACING_SIZES: SpacingSize[] = ['xs', 'sm', 'df', 'md', 'lg', 'xl'];
+
+export const TEXT_SIZES: TextSize[] = ['sm', 'df', 'lg'];
+
+export const PALETTE_TYPES: PaletteType[] = [
+  'brand',
+  'primary',
+  'secondary',
+  'tertiary',
+  'neutral',
+  'muted',
+  'danger',
+  'warning',
+  'failure',
+  'success',
+  'info',
+];
+
+export const SHADE_RANGES: ColorShade[] = [
+  '00',
+  '10',
+  '20',
+  '30',
+  '40',
+  '50',
+  '60',
+  '70',
+  '80',
+  '90',
+];
